restructured functional tests
